fix rds snapshot purging
The list_old_rds_snapshot_ids function did not filter out snapshots
in an unavailable state.  Thus, if you had KEEP_BACKUPS set to a number
greater than the current existing number of backups (ie: the first few
times you attempted to make a backup), it would include the just-now-
created snapshot in the list of snapshots to purge, and the script
would fail with this error:

```
An error occurred (InvalidDBSnapshotState) when calling the
DeleteDBSnapshot operation: Cannot delete the snapshot because it is not
currently in the available or failed state.
```

This PR uses the awscli jmespath query function to filter only for
snapshots with the "available" status.  Since using two json parsers
in the same command line is silly, the rest of the jq logic is
also replicated in jmespath.