BSERV-12859 The Backup & Restore using Bitbucket DIY script for Mysql DB fails for below reasons:
BITBUCKET_BACKUP_DB variable is a directory instead of file.
The mysqldump command would throw 'Access Denied' error if the bitbucketuser credentials is used in the config file.
The restore would fail if the user hasn't created the database before executing the restore script.

This fix would resolve the above issue for Bitbucket MYSQL instance